commit 3

+++ Population.cpp
@@ -20,7 +20,7 @@ Population::Population(int size, int MaxDur, double prob_cross) {
 void Population::new_generation() {
     std::vector<Chromosome> new_generation{static_cast<size_t>(population_size)};
 
-    int n_percent = population_size * 0.1;
+    int n_percent = static_cast<int>(population_size * 0.1);
     std::sort(population_array.begin(), population_array.end(),
               [](Chromosome x, Chromosome y) { return Chromosome::cost_function(x) > Chromosome::cost_function(y); });
 
@@ -44,8 +44,7 @@ void Population::new_generation() {
         while (parent_one_index == parent_two_index) {
             parent_two_index = dist(rng);
         }
-        new_generation[i] = Chromosome::cross(this->population_array[parent_one_index],
-                                              this->population_array[parent_two_index]);
+        new_generation[i] = Chromosome::cross(this->population_array[parent_one_index], this->population_array[parent_two_index]);
     }
     std::cout << "Crossover process completed.\n";
 
@@ -53,8 +52,7 @@ void Population::new_generation() {
     // one or more new members of the population are randomly generated
     // from the same distribution as the original population
     for (; i < population_size; ++i) {
-        new_generation.emplace_back(
-                Chromosome(population_size, new_generation[0].get_max_dur(), new_generation[0].get_prob_cross()));
+        new_generation.emplace_back(Chromosome(population_size, new_generation[0].get_max_dur(), new_generation[0].get_prob_cross()));
     }
     this->population_array.clear();
     this->population_array = new_generation;
